refactored specificatin classes to a subpackage under repositories
