+++ src/main/java/com/nfyc/studyplanservice/services/impl/StudyPlanServiceImpl.java
@@ -3,25 +3,17 @@ package com.nfyc.studyplanservice.services.impl;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.nfyc.studyplanservice.mappers.CourseMapper;
 import com.nfyc.studyplanservice.mappers.TopicMapper;
-import com.nfyc.studyplanservice.model.domain.Course;
-import com.nfyc.studyplanservice.model.domain.CourseSpecifications;
-import com.nfyc.studyplanservice.model.domain.FILTER_OPERATORS;
-import com.nfyc.studyplanservice.model.dto.FilterRequestDTO;
+import com.nfyc.studyplanservice.repositories.specifications.CourseSpecifications;
 import com.nfyc.studyplanservice.model.dto.FilterRequestListDTO;
 import com.nfyc.studyplanservice.model.dto.StudyPlanDTO;
 import com.nfyc.studyplanservice.model.dto.StudyPlanListDTO;
 import com.nfyc.studyplanservice.repositories.CourseRepository;
-import com.nfyc.studyplanservice.repositories.TopicRepository;
-import com.nfyc.studyplanservice.services.CourseService;
 import com.nfyc.studyplanservice.services.StudyPlanService;
-import com.nfyc.studyplanservice.services.TopicService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import javax.validation.ConstraintViolation;
